+++ apps/web/src/app/app-routing.module.ts
@@ -1,11 +1,16 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
+import { HomeComponent } from './home/home.component';
 
 const routes: Routes = [
   {
     path: 'escolas',
     loadChildren: () =>
       import('./schools/schools.module').then((m) => m.SchoolsModule),
+  },
+  {
+    path: '**',
+    component: HomeComponent,
   },
 ];